+++ examples/setup-temp/routes/index.tsx
@@ -5,7 +5,7 @@ import Counter from "../islands/Counter.tsx";
 export default function Home() {
   const count = useSignal(3);
   return (
-    <Main>
+    <Main whitespace>
       <div class="max-w-screen-md mx-auto flex flex-col items-center justify-center">
         <img
           class="my-6"